refactor: 병원 정보 가져오는 기능 클래스 분리
